Add .env, client structure.

+++ src/main/java/com/denis/java_test_client/models/Client.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class Client {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer client_id;  //идентификатор клиента
+    private Long client_id;  //идентификатор клиента
     @Column(nullable = false)
     private String name;        //имя клиента
     @Column(nullable = false)
@@ -21,7 +21,7 @@ public class Client {
     @OneToOne @JoinColumn(name = "сontact_id")
     private Contact contact_id; //контакты клиента
 
-    public Client(Integer client_id, String name, String last_name) {
+    public Client(Long client_id, String name, String last_name) {
         this.client_id = client_id;
         this.name = name;
         this.last_name = last_name;
